fix(EMS-2009): dry api test helper
