+++ src/api/generate-xlsx/map-application-to-XLSX/index.test.ts
@@ -7,24 +7,14 @@ import mapPolicyAndExport from './map-policy-and-export';
 import mapExporter from './map-exporter';
 import mapBuyer from './map-buyer';
 import mapEligibility from './map-eligibility';
-import getPopulatedApplication from '../../helpers/get-populated-application';
-import { createFullApplication, getKeystoneContext, mapApplicationIds } from '../../test-helpers';
+import { generateSubmittedApplication } from '../../test-helpers';
 import { Application } from '../../types';
 
 describe('api/generate-xlsx/map-application-to-xlsx/index', () => {
   let submittedApplication: Application;
 
   beforeEach(async () => {
-    const context = getKeystoneContext();
-    const application = await createFullApplication(context);
-
-    const applicationIds = mapApplicationIds(application);
-    const populatedApplication = await getPopulatedApplication(context, applicationIds);
-
-    submittedApplication = {
-      ...populatedApplication,
-      submissionDate: new Date(),
-    };
+    submittedApplication = await generateSubmittedApplication();
   });
 
   it('should return an array of mappings and section breaks', () => {
