+++ src/api/generate-xlsx/styled-columns/index.test.ts
@@ -3,8 +3,7 @@ import styledColumns, { worksheetRowHeights } from '.';
 import HEADER_COLUMNS from '../header-columns';
 import mapApplicationToXLSX from '../map-application-to-XLSX';
 import { XLSX_CONFIG, XLSX_ROW_INDEXES } from '../../constants';
-import getPopulatedApplication from '../../helpers/get-populated-application';
-import { createFullApplication, getKeystoneContext, mapApplicationIds } from '../../test-helpers';
+import { generateSubmittedApplication } from '../../test-helpers';
 import { mockApplication } from '../../test-mocks';
 
 const { LARGE_ADDITIONAL_COLUMN_HEIGHT, ADDITIONAL_TITLE_COLUMN_HEIGHT, FONT_SIZE } = XLSX_CONFIG;
@@ -79,16 +78,7 @@ describe('api/generate-xlsx/styled-columns/index', () => {
 
   describe('worksheetRowHeights', () => {
     it('should add column heights to particular columns', async () => {
-      const context = getKeystoneContext();
-      const application = await createFullApplication(context);
-
-      const applicationIds = mapApplicationIds(application);
-      const populatedApplication = await getPopulatedApplication(context, applicationIds);
-
-      const submittedApplication = {
-        ...populatedApplication,
-        submissionDate: new Date(),
-      };
+      const submittedApplication = await generateSubmittedApplication();
 
       const xlsxData = mapApplicationToXLSX(submittedApplication);
