+++ src/api/test-helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './create-full-application';
 export * from './get-keystone-context';
 export * from './map-application-ids';
+export * from './generate-submitted-application';
